python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meetingscribe"
version = "0.1.0"
description = "Auto-detect meetings, transcribe, summarize, output to Obsidian"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.1.0",
    "pyannote-audio>=3.3",
    "watchdog>=4.0.0",
    "keyring>=25.0.0",
    "tomli>=2.0.0;python_version<'3.12'",
    "tomli-w>=1.0.0",
    "numpy>=1.26.0",
    "soundfile>=0.12.0",
    "scipy>=1.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-tmp-files>=0.0.2"]

[project.scripts]
meetingscribe = "meetingscribe.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/meetingscribe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
